qa/requirements.txt
django-allauth[socialaccount]  # social login
django-silk  # performance profile
django-simple-history  # to keep track of the history of model changes
psycopg  # postgres connection
PyYaml  # format fixtures in YAML
